feat: Allow LocationId based agency resolution in Polaris whilst retaining the existing behaviour as a default DCB-1675
Adds the optional capability for the Polaris client to use LocationId per DCB-1675 as a strategy for resolving an item agency whilst retaining the current code strategy as a default.

Notes:

Behaviour is set via the HostLMS setting clientConfig.item.item-agency-resolution-method which defaults to "Legacy" (The current behaviour).  Correct function of LocationId strategy depends upon the presence of mappings (direct_mappings.tsv) which map LocationId as a string to agency. For exmaple:

    fromContext fromCategory  fromValue toContext toCategory  toValue
    RIVERSHARE  Location  16  DCB AGENCY  6davn

It is crucially important to remember that when resolving an agency as a part of the Search/RTAC flow the HostLMS adapter used will be the "Catalog" role. For
instances of locate based on a shared system, this will be the HostLMS of the "Shared" system, and NOT the "Circulation" LMS for each library. Note also that
the fromContext in this scenario is the shared system not the target library - as we are trying to derive from a shared system which of the member libraries
a given item belongs to.

When LocationID agency resolution is enabled the concept of "Default Agency" for Polaris systems is not used when mapping items - we either know the agency explicitly or we don't and whilst we will return information about the item, we cannot use it in subsequent flows.