feat: Implement on-site borrowing [DCB-1555]
Introduces a new ExpeditedCheckoutTransition to facilitate on-site-borrowing. This transition ensures that once patron requests flagged with 'isExpeditedCheckout' reach REQUEST_PLACED_AT_BORROWING_AGENCY, they speed through to LOANED so they can be immediately checked out at the supplier.

This is useful for any other scenario where we do not want to follow the normal tracking cycle and we need an 'instant checkout'. Practically checkout will not be completely instant, but it should be <1m.