chore: Improve data change log recording for library entities[DCB-1744]
Adds a migration to ensure that changes to consortium contacts are logged (particularly deletes). Also improves data change log entries for adding a library to a group, adding or updating functional settings, deleting a contact for either a library or a consortia, and initial creation of contacts for libraries and consortia.
This should ensure that all library entity operation entries in the data change log have a username, a reason and a change category.