fix: Prevent all source records from being set to PROCESSING_REQUIRED [DCB-1886]
The processingState of all source records has been showing as PROCESSING_REQUIRED in DCB Admin, but shows up as expected in the database. I believe this is because the processing state was being defaulted to PROCESSING_REQUIRED in the SourceRecord class: because processingState is declared as a final, this means that Micronaut Data cannot populate it with the correct value from the database. Hence, all SourceRecords' processing states default to PROCESSING_REQUIRED and never reflect the changes from the DB.