Prevent Alma adaptor wrongly trying to create a new user
This is an experimental fix, so I have deliberately not used conventional commit as I don't want this or other previous experimental commits to clog up the release logs. These experiments are being committed as my local environment's Alma setup fails because it is unauthorised, and there is significant time pressure on getting a successful Alma request.