fix: Prevent location mapping creation failure [DCB-1937]
Location mapping creation was not allowing the creation of any location mapping with a fromValue that already existed for any library, when it should only have been disallowing creation for fromValues that existed at the current library.

This was happening because the toContext and category are always the same for location mappings - unlike other mappings - and the fix is to ensure that this check is not applied for Location mappings, as it will always fail.