fix: Improve validation of location mappings [DCB-1938]
Disallows the creation of location mappings with a from and to category of "Location", and enforces the fromCategory of "Location" and toCategory of "AGENCY" to ensure all created location mappings are valid. This should prevent a reoccurrence of the issue in DCB-1938 where a large number of location mappings had been created incorrectly, causing issues displaying items.