feat: Implement supplying agency filter [DCB-1973]
To eliminate a particularly unwieldy and unreliable work-around in DCB Admin and Admin for Libraries, I have implemented a special use case in our server-side filtering support for the supplying agency code. This uses a join to enable us to provide an "equals" filter for supplying agency code in our frontend applications.

This is particularly important for DCB Admin for Libraries as this filter is key to providing the "library supplier requests" view - the existing frontend work-arounds fail when dealing with supplier request libraries such as UMSYS with many thousands of requests.