chore: Second migration to catch nulls from 8.45 [DCB-2001]
Because of the missing @Lombok.Builder annotation on isTooLong in my previous change from 8.45, patron requests constructed with the builder will have seen null overwrite their default false value. This has led to behaviour where they appear to have "false" set when querying with GraphQL, but actually have null set in the DB. As such, a second migration is necessary to catch the patron requests on 8.45 systems which have isTooLong values of null.