fix: Use kebab-case so active-request-limit is configurable [DCB-2007]
When attempting to configure the activeRequestLimit, Micronaut did not seem to be able to match the env variables from scripts (DCB_GLOBALS_ACTIVE_REQUEST_LIMIT) and so the active request limit remained at 50. Changing activeRequestLimit to kebab-case makes it configurable via scripts, brings it in line with other env variables such as the polling durations, and will make it easier for external partners such as FSE to understand. I have also updated all references to the previous value.