fix: Prevent renewal prevention from trying to cancel the original loan [DCB-2006]
Because of idiosyncrasies in edge and mod-dcb, DCB is attempting to prevent renewal on FOLIO requests that make it to LOANED, as it is receiving renewable: false without any other information. While we wait for changes in those modules, this fix detects the initial loan (when all known renewal counts are zero) and ensures that the HandleSupplierHoldDetected does not trigger renewal prevention in these circumstances. This restores the ability to progress FOLIO DCB requests past LOANED. More work in this area and the associated modules will be required to strengthen renewal prevention without affecting the request workflow.