feat: Suppression in folio sources and default rules for those
Wired in suppression rules to folio. Added collection iteration for
finding array items with a particular property to allow for the format
of the source here where each property is in an object with a single
property. Refactored out the blocking code for this source too which
opens up the path to repeat for the others.

default suppression rules for folio found in config.

provides: [DCB-1701]