fix: Get the correct renewable value from Polaris [DCB-2022]
Previously, we were not getting the actual renewable value from the Polaris item. This led to DCB constantly thinking the supplier item's renewable status had changed (for Polaris suppliers) and constantly recording this in the audit log. This was because there was a conflict between the renewable and 'local renewable' values that never got resolved, as we never updated the item with the actual renewable value.