feat: Update Polaris virtual patron expiry [DCB-2059]
Introduces new methods to update the expiry date of a Polaris virtual patron that is either about to expire (7 days) or has already expired.
This will activate the next time DCB looks up the virtual patron in question, ensuring virtual patrons are updated before they are used and preventing request failure as a result of virtual patrons becoming blocked when they expire.