fix: Update Polaris virtual patron expiry check cut-off to 30 days [DCB-2069]
After feedback from MCO, we will now update a Polaris virtual patron's expiry date if it is within 30 days of expiring. We will also extend the expiry date by 120 days instead of 90.
This change also removes a legacy version of this implementation that was left in the PolarisLmsClient by mistake.