fix: Fix Polaris cleanup behaviour [DCB-2076]
Previously, cleanup would report a success even if it failed to delete a virtual item or virtual request for Polaris. The presence of a hold request on a virtual item in Polaris will mean that we always fail to delete the virtual item when we try to do so.

Now, we will first attempt to delete the hold on the virtual item, but if that fails we will resort to cancelling it. This gets rid of the hold and will prevent this from causing virtual item deletion to fail.

The end result is that now both the virtual request and the virtual item are correctly cleaned up. This should also at least mitigate the issue with virtual item creation failing because of duplicate barcodes, although manual cleanup in existing systems may be required.