fix: Handle more complicated cleanup scenarios [DCB-2076]
As per the Polaris docs, there are situations where the workflow will succeed but the item won't necessarily be deleted. These changes improve our handling of the different potential scenario and harden the workflow as Polaris does not always give us a standard success response - indeed it can actually delete the item and give a range of responses.