fix: Make out of sequence consider all valid workflows [DCB-2085]
Previously, any request that did not follow the RET-STD path was considered out of sequence. This has caused issues with perfectly legitimate requests being flagged as potentially problematic. It also undermines confidence in the workflow.

Now, out of sequence will only apply to a request when it actually deviates from any valid workflow.