fix: De-prioritise overdue CHECKED_OUT items [DCB-2075]
An issue was reported where DCB was resolving to overdue CHECKED_OUT items ahead of AVAILABLE ones. This was because DCB took the dueDate from the local systems (which was some time in the past) on trust, and used it directly for the availability date. As a result the most overdue CHECKED_OUT item was being selected because it was seen as having the earliest availability date - despite this being in some cases a significant time in the past, without ever becoming AVAILABLE.

This change alters this behaviour so that we treat the due date for overdue CHECKED_OUT items as untrustworthy, and by default give them an availability date of the current date + the default loan period (28 days), multiplied by the hold count + 1 (so that a CHECKED_OUT item with a hold count of zero will have an availability date of 28 days from the present date). As such, the fallback behaviour is preserved - DCB will select the earliest available CHECKED_OUT item if no items are AVAILABLE - but in practice DCB will never prioritise a CHECKED_OUT item over an AVAILABLE one.