fix: Use the correct local ID when deleting a supplier hold [DCB-2068]
For Polaris systems, the existing supplier patron ID that was being used when deleting a hold can be null. This causes the deletion to fail silently. This change ensures that if supplierPatronId is null, we grab the virtual patron's local ID and use that instead, so we can successfully delete the hold.

This change deliberately leaves the supplierPatronId as the first point of call, and will only kick in if that is a null value.