fix: Improve Polaris hold deletion and audit logging [DCB-2068]
Improves our Polaris hold deletion approach. Now if we fail to delete a hold originally, we fall back to cancelling it and then try and delete it again if we are successful at cancelling it.

This aims to fix the problem whereby we can't delete a hold if it is in a certain status. Currently, we just cancel it and leave it at that - this goes further and makes sure we at least try and delete it after cancelling it.

These changes also improve audit logging around deleting supplier holds. Now we should get an audit entry for the different potential success outcomes.