fix: Get Polaris hold count if reservation data not present [DCB-2093]
When reservation data is not available, we may still be able to get the hold count from the "total count" on the response. Failing to do this is causing us to incorrectly report zero holds on some Polaris items. Which can then lead to the "duplicate holds" error when we resolve to an item which has holds that DCB is not aware of.