feat: Prevent renewals [DCB-1533]
Prevent renewals when we detect a hold on the requested item in the local system. This implementation makes sure to check the up-to-date hold count before preventing the renewal, to ensure we don't accidentally block renewals based on out of date data (i.e. if we've not registered that our hold is cleared yet)

This also adds audits to be clear about which stage of renewal prevention has been achieved, and if we have successfully avoided a false positive.