fix: Correctly determine FOLIO hold counts [DCB-1533]
As part of the prevent renewals work, this change makes our method of determining the hold count more robust.
Depending on what we get back from mod-dcb, the hold count can be directly on the transaction status OR it can be on the item object on the transaction status.

If no hold count is received, we default to 0.