fix: Correctly determine Polaris hold counts [DCB-2093]
Previously, our Polaris hold count values in DCB were always 0. This was because we were not passing a Limit parameter, and so the Polaris API would not give us the reservation data.

We now obtain this data by passing the Limit parameter, filter it for cancelled holds (which Polaris counts, but we're not interested in), and then filter again to ensure we are only including item-level holds.

We also now place a paginated request for the item holds, to improve our handling of situations where an item may have many holds.

If we would like to also incorporate bib level holds in the hold count, we can greatly reduce the complexity required here.