feat: Include real names in Polaris virtual patron records [DCB-2058]
Adds the functionality for including real names in Polaris virtual patron records if the VIRTUAL_PATRON_NAMES_POLARIS setting is enabled.

Also ensures that the unique identifiers are still included, typically on the last name, for both Sierra and Polaris, as these are currently used by librarians and will be useful for identifying virtual patrons with common names.