fix: Make sure we fail safely if the prevent renewal endpoint is not available [DCB-1533]
If the prevent renewal endpoint is not available, or renewal prevention otherwise fails, fail safely and do not put the request into error.
Also makes an audit log entry in this situation, and guards the Polaris handling of complicated hold counts: if the total is 0, we return 0 and don't make extra requests.