fix: Fix resolution selecting item with holds over available one [DCB-2106]
It appears that in certain situations we have been selecting the closest geographical item even if it's not the earliest available. I suspect that we are entering a tie breaker situation possibly incorrectly and applying geo sort to the whole list as a result (when it should just be the real top items).